requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-xml==2.0.0
Faker==25.0.1
Markdown==3.6
pillow==10.3.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
validate-docbr==1.10.0
